pyproject.toml
[project]
name = "dp100-monitor"
version = "0.1.0"
description = "Real-time power monitoring for Alientek DP100"
authors = [
    {name = "bergmann"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dash (>=3.2.0,<4.0.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "h5py (>=3.15.1,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "hidapi (>=0.14.0.post4,<0.15.0)",
    "dash-bootstrap-components (>=2.0.4,<3.0.0)",
    "psutil (>=7.1.2,<8.0.0)",
    "crcmod (>=1.7,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"

